Hehe

+++ src/comments/comment-use-cases/delete-comment-use-case.ts
@@ -2,9 +2,14 @@ import { Inject } from '@nestjs/common';
 import { CommentsService } from '../comments.service';
 import { CommentsRepository } from '../repositories/comments.repository';
 import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
+import { DomainException } from '../../core/exceptions/domain-exceptions';
+import { DomainExceptionCode } from '../../core/exceptions/domain-exception-codes';
 
 export class DeleteCommentCommand {
-  constructor(public commentId: string) {}
+  constructor(
+    public commentId: string,
+    public userId: string,
+  ) {}
 }
 
 @CommandHandler(DeleteCommentCommand)
@@ -14,7 +19,16 @@ export class DeleteCommentUseCase implements ICommandHandler<DeleteCommentComman
     @Inject(CommentsRepository) private commentsRepository: CommentsRepository,
   ) {}
   async execute(command: DeleteCommentCommand): Promise<void> {
-    await this.commentsService.findCommentById(command.commentId);
+    const foundComment = await this.commentsService.findCommentById(
+      command.commentId,
+    );
+    if (foundComment.commentatorInfo.userId !== command.userId) {
+      throw new DomainException({
+        code: DomainExceptionCode.Forbidden,
+        field: 'jwtToken',
+        message: 'You dont have permission to delete a comment',
+      });
+    }
     await this.commentsRepository.deleteComment(command.commentId);
   }
 }
